+++ CHANGELOG.md
@@ -1,5 +1,25 @@
 # Changelog
 
+## [0.3.1](https://github.com/runpod-workers/worker-tetra/compare/v0.3.0...v0.3.1) (2025-07-23)
+
+
+### Bug Fixes
+
+* broken ci ([#13](https://github.com/runpod-workers/worker-tetra/issues/13)) ([b25d822](https://github.com/runpod-workers/worker-tetra/commit/b25d8220ef0389dea6a83fd9a4450be459e79244))
+
+## [0.3.0](https://github.com/runpod-workers/worker-tetra/compare/v0.2.0...v0.3.0) (2025-07-23)
+
+
+### Features
+
+* AE-835 Add class based execution [Runtime] ([#8](https://github.com/runpod-workers/worker-tetra/issues/8)) ([6d6505e](https://github.com/runpod-workers/worker-tetra/commit/6d6505ebdd749dff45dd52cb18b93da9330fe5ab))
+* CI/CD pipeline workflows with testing, linting, valiation and docker builds ([#9](https://github.com/runpod-workers/worker-tetra/issues/9)) ([9d3d696](https://github.com/runpod-workers/worker-tetra/commit/9d3d69698238718ab64675b335630caf3c186526))
+
+
+### Bug Fixes
+
+* update Dockerfile to reference only existing files ([#12](https://github.com/runpod-workers/worker-tetra/issues/12)) ([93df475](https://github.com/runpod-workers/worker-tetra/commit/93df4756bea1c60adae9063cd2426ea230f3b7d5))
+
 ## [0.2.0](https://github.com/runpod-workers/worker-tetra/compare/v0.1.1...v0.2.0) (2025-06-26)